Renamed ExtensionRepoApi.kt to ExtensionRepoService.kt
