+++ domain/src/main/java/mihon/domain/extensionrepo/interactor/UpdateExtensionRepo.kt
@@ -4,16 +4,16 @@ import eu.kanade.tachiyomi.network.NetworkHelper
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
-import mihon.domain.extensionrepo.api.ExtensionRepoApi
 import mihon.domain.extensionrepo.model.ExtensionRepo
 import mihon.domain.extensionrepo.repository.ExtensionRepoRepository
+import mihon.domain.extensionrepo.service.ExtensionRepoService
 
 class UpdateExtensionRepo(
     private val extensionRepoRepository: ExtensionRepoRepository,
     networkService: NetworkHelper,
 ) {
 
-    private val extensionRepoApi = ExtensionRepoApi(networkService.client)
+    private val extensionRepoService = ExtensionRepoService(networkService.client)
 
     suspend fun awaitAll() = coroutineScope {
         extensionRepoRepository.getAll()
@@ -22,7 +22,7 @@ class UpdateExtensionRepo(
     }
 
     suspend fun await(repo: ExtensionRepo) {
-        val newRepo = extensionRepoApi.fetchRepoDetails(repo.baseUrl) ?: return
+        val newRepo = extensionRepoService.fetchRepoDetails(repo.baseUrl) ?: return
         if (
             repo.signingKeyFingerprint.startsWith("NOFINGERPRINT") ||
             repo.signingKeyFingerprint == newRepo.signingKeyFingerprint
